task 5.Products service refactoring
