+++ common/types/serviceResponse.type.ts
@@ -1,4 +1,5 @@
 export type ServiceResponse = {
   statusCode: number;
   body: any;
+  headers?: Record<string, unknown>;
 }
