+++ common/utils/index.ts
@@ -1,12 +1,11 @@
-import { APIGatewayProxyResult } from "aws-lambda";
 import { ServiceResponse } from "../types/serviceResponse.type";
 
-export function mapToProxyResult(data: ServiceResponse): APIGatewayProxyResult {
+export function mapToProxyResult(data: ServiceResponse): ServiceResponse {
   return {
     statusCode: data.statusCode,
     headers: {
       'Access-Control-Allow-Origin': '*',
-      'Access-Control-Allow-Credentials': true,
+      'Access-Control-Allow-Credentials': false,
     },
     body: JSON.stringify(data.body, null, 2)
   }
